refactor: rename bike point search delegate
